refactor error classes
