+++ classes/DataValidationError.js
@@ -1,0 +1,13 @@
+const { ERRORS, MESSAGES } = require('../utils/constants');
+
+class DataValidationError extends Error {
+  constructor(message = MESSAGES.uncorrectData) {
+    super(message);
+    this.name = 'DataValidationError';
+    this.statusCode = ERRORS.UNCORRECT;
+  }
+}
+
+module.exports = {
+  DataValidationError,
+};
